@@ -21,7 +21,7 @@ apply plugin:'application'
 
 mainClassName = "org.apache.bigtop.bigpetstore.qstream.LoadGen"
 
-sourceCompatibility = 1.7
+sourceCompatibility = 1.8
 version = '1.0'
 
 manifest {
@@ -41,22 +41,15 @@ task fatJar(type: Jar) {
 
 
 repositories {
+    mavenLocal()
     mavenCentral()
-    maven {
-       url "https://dl.bintray.com/rnowling/bigpetstore"
-   }
 }
 
 dependencies {
-    compile('com.fasterxml:jackson-module-json-org:0.9.1') {
-        exclude group: 'org.apache.commons', module: 'commons-io'
-    }
-    compile('org.apache.commons:commons-io:1.3.2') {
-         exclude group: 'org.apache.commons', module: 'commons-io'
-         exclude group: 'commons-io', module: 'commons-io'
-    }
-    compile 'org.apache.httpcomponents:httpclient:4.4'
-    compile group:'org.apache.commons', name:'commons-lang3',version:'3.3.2'
+    compile 'com.fasterxml.jackson.core:jackson-databind:2.18.2'
+    compile 'commons-io:commons-io:2.18.0'
+    compile 'org.apache.httpcomponents:httpclient:4.5.14'
+    compile 'org.apache.commons:commons-lang3:3.17.0'
     compile "com.github.rnowling.bigpetstore:bigpetstore-data-generator:0.2.1"
-    testCompile group: 'junit', name: 'junit', version: '4.11'
+    testCompile group: 'junit', name: 'junit', version: '4.13.2'
 }
